Extract dependency graph functions into carina-core::deps module
Move resource dependency analysis functions from carina-cli/src/main.rs
to a new carina-core/src/deps.rs module for better testability and reuse.
Also add resource() and binding_name() methods to Effect.

Refs #328
